added schemaHelpers.ts file
