+++ packages/postman-to-openapi/src/helpers/responseHelpers.ts
@@ -1,6 +1,7 @@
 import type { OpenAPIV3 } from '@scalar/openapi-types'
 
 import type { HeaderList, Response } from '../types'
+import { inferSchemaFromExample } from './schemaHelpers'
 
 /**
  * Extracts and converts Postman response objects to OpenAPI response objects.
@@ -22,44 +23,13 @@ export function extractResponses(
       headers: extractHeaders(response.header),
       content: {
         'application/json': {
-          schema: inferResponseSchema(response.body || ''),
+          schema: inferSchemaFromExample(response.body || ''),
           example: tryParseJson(response.body || ''),
         },
       },
     }
     return openapiResponses
   }, {} as OpenAPIV3.ResponsesObject)
-}
-
-function inferResponseSchema(body: string): OpenAPIV3.SchemaObject {
-  try {
-    const parsedBody = JSON.parse(body)
-    return inferSchemaFromExample(parsedBody)
-  } catch (e) {
-    return { type: 'string' }
-  }
-}
-
-function inferSchemaFromExample(example: any): OpenAPIV3.SchemaObject {
-  if (Array.isArray(example)) {
-    return {
-      type: 'array',
-      items: example.length > 0 ? inferSchemaFromExample(example[0]) : {},
-    }
-  } else if (typeof example === 'object' && example !== null) {
-    const properties: { [key: string]: OpenAPIV3.SchemaObject } = {}
-    for (const [key, value] of Object.entries(example)) {
-      properties[key] = inferSchemaFromExample(value)
-    }
-    return {
-      type: 'object',
-      properties,
-    }
-  } else {
-    return {
-      type: typeof example as OpenAPIV3.NonArraySchemaObjectType,
-    }
-  }
 }
 
 function extractHeaders(
